Move build info constants into their own module
